+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## Unreleased
+### Added
+- Allow to override base docker images
+
+### Modified
+- Slim image for openjdk/jdk17
 
 ## 0.1.23
 ### Added